requirements.txt
flask
numpy
matplotlib
pillow
